test: Add tests for config, cache and api modules (#41)
* add: Add pytest config

Include pytest config into `pyproject.toml`

* add: Add suite of test for cache module

* chore: Correct misleading comment

A comment was not technically correct so it could lead to
misunderstading in the method functionality.

* test: Add suite of tests for config module (#39)

* refactor: Receive path as parameter

Set path as a function parameter in `init_config_file()`. This will help
if in the future its decided to set customs config file paths.

* test: Add suite of tests for config module

* Add suite of test for API module (#40)

* refactor: Use verify param

Change param from `AkamaiAPI._validate_certs` to `AkamaiAPI._session_verify`
to match with the parameters names in `requests.Session`.

Also `_post` and `_patch` method receive `json` instead of `payload`.

* fix: Remove payload from GET request

The function was passing `_poll_if_needed` was passing json param to
second and later tries, which always use GET method, so it would fail.

* docs: Remove unneeded comment

* test: Add suite of test for api module